Move all private utils.py function to utils_helper.py
