+++ accountlockout/ip.py
@@ -1,5 +1,6 @@
 from defender import config as def_config
 
+import accountlockout.helper.utils_helper
 from accountlockout.helper.ip_helper import REDIS_SERVER, __get_blocked_cache_key
 from . import utils
 
@@ -12,4 +13,4 @@ def get_blocked_ips():
     key = __get_blocked_cache_key("*")
     key_list = [redis_key.decode('utf-8')
                 for redis_key in REDIS_SERVER.keys(key)]
-    return utils.__strip_keys(key_list)
+    return accountlockout.helper.utils_helper.__strip_keys(key_list)
