+++ accountlockout/users.py
@@ -1,6 +1,7 @@
 from defender import config as def_config
 
 import accountlockout.helper.ip_helper
+import accountlockout.helper.utils_helper
 from accountlockout.helper.users_helper import REDIS_SERVER, __get_username, __lower, __get_attempt_cache_key, \
     __get_blocked_cache_key
 from . import utils
@@ -43,5 +44,5 @@ def get_blocked_usernames():
     key = __get_blocked_cache_key("*")
     key_list = [redis_key.decode('utf-8')
                 for redis_key in REDIS_SERVER.keys(key)]
-    return utils.__strip_keys(key_list)
+    return accountlockout.helper.utils_helper.__strip_keys(key_list)
